chore: organize test models
