+++ tests/Feature/RevisionScopesTest.php
@@ -3,9 +3,8 @@
 namespace Plank\Checkpoint\Tests\Feature;
 
 use Plank\Checkpoint\Models\Checkpoint;
-use Plank\Checkpoint\Models\Revision;
 use Plank\Checkpoint\Scopes\RevisionScope;
-use Plank\Checkpoint\Tests\Support\Post;
+use Plank\Checkpoint\Tests\Support\Models\Blog\Post;
 use Plank\Checkpoint\Tests\TestCase;
 
 class RevisionScopesTest extends TestCase
