+++ tests/Feature/RevisionableRelationsTest.php
@@ -4,7 +4,7 @@ namespace Plank\Checkpoint\Tests\Feature;
 
 use Plank\Checkpoint\Models\Checkpoint;
 use Plank\Checkpoint\Models\Revision;
-use Plank\Checkpoint\Tests\Support\Post;
+use Plank\Checkpoint\Tests\Support\Models\Blog\Post;
 use Plank\Checkpoint\Tests\TestCase;
 
 class RevisionableRelationsTest extends TestCase
